feat: create builderv2
Start a new implementation of the builder following the builder pattern. This new builder will create dialogue graphs in the form of entity nodes instead of a petgraph component. As of this commit there is a simple say method to add text nodes and a build method that creates a bevy Command to be applied to the world.